New architecture for reducers
Instead of `compose` tracking and passing accumulators and values
it now only passes the next value to each reducer. It is up to
the reducer to do its own accumulator collecting.

In order for `compose` to get the accumulator, it passes a new
`isFinal` argument to each reducer. When `isFinal` is `true` the
reducer is the final one in the composed reducer chain, and
should return the accumulator. When `isFinal` is `false` the
reducer should return the value to be passed to the next reducer.

This should improve memory consumption when multiple reducers are
composed together because only the final reducer in the chain
actually builds up an accumulator.

This new design also makes it so `compose` doesn't need to know
about the `initial` value of the reducer, so reducer creators
can just return a plain reducer function.