Fix an edge case with passing `undefined` to reducers
Previously `undefined` was used as a marker value to indicate
that processing with reducers should short circuit (return
without further processing). This is used by the `filter` reducer.

In case someone has a reason to return `undefined` from a reducer
and pass that on to additional reducers, the code now uses a
Symbol as the marker constant indicating that reducer processing
should end immediately.